+++ src/main/java/encore/server/global/auth/UserDetailsImpl.java
@@ -1,0 +1,30 @@
+package encore.server.global.auth;
+
+import java.util.Collection;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.userdetails.UserDetails;
+
+@Getter
+@AllArgsConstructor
+public class UserDetailsImpl implements UserDetails {
+
+  private Long userId;
+  private String email;
+
+  @Override
+  public Collection<? extends GrantedAuthority> getAuthorities() {
+    return null;
+  }
+
+  @Override
+  public String getPassword() {
+    return null;
+  }
+
+  @Override
+  public String getUsername() {
+    return null;
+  }
+}
